@@ -2,12 +2,18 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import frequently from '../utils/frequently'
+import { getData } from '../utils'
 import { Emoji } from '.'
 
 export default class Category extends React.Component {
+  constructor(props) {
+    super(props)
+
+    this.setContainerRef = this.setContainerRef.bind(this)
+    this.setLabelRef = this.setLabelRef.bind(this)
+  }
+
   componentDidMount() {
-    this.container = this.refs.container
-    this.label = this.refs.label
     this.parent = this.container.parentNode
 
     this.margin = 0
@@ -17,11 +23,24 @@ export default class Category extends React.Component {
   }
 
   shouldComponentUpdate(nextProps, nextState) {
-    var { name, perLine, native, hasStickyPosition, emojis, emojiProps } = this.props,
-        { skin, size, set } = emojiProps,
-        { perLine: nextPerLine, native: nextNative, hasStickyPosition: nextHasStickyPosition, emojis: nextEmojis, emojiProps: nextEmojiProps } = nextProps,
-        { skin: nextSkin, size: nextSize, set: nextSet } = nextEmojiProps,
-        shouldUpdate = false
+    var {
+        name,
+        perLine,
+        native,
+        hasStickyPosition,
+        emojis,
+        emojiProps,
+      } = this.props,
+      { skin, size, set } = emojiProps,
+      {
+        perLine: nextPerLine,
+        native: nextNative,
+        hasStickyPosition: nextHasStickyPosition,
+        emojis: nextEmojis,
+        emojiProps: nextEmojiProps,
+      } = nextProps,
+      { skin: nextSkin, size: nextSize, set: nextSet } = nextEmojiProps,
+      shouldUpdate = false
 
     if (name == 'Recent' && perLine != nextPerLine) {
       shouldUpdate = true
@@ -31,7 +50,13 @@ export default class Category extends React.Component {
       shouldUpdate = !(emojis == nextEmojis)
     }
 
-    if (skin != nextSkin || size != nextSize || native != nextNative || set != nextSet || hasStickyPosition != nextHasStickyPosition) {
+    if (
+      skin != nextSkin ||
+      size != nextSize ||
+      native != nextNative ||
+      set != nextSet ||
+      hasStickyPosition != nextHasStickyPosition
+    ) {
       shouldUpdate = true
     }
 
@@ -58,7 +83,6 @@ export default class Category extends React.Component {
     margin = margin > this.maxMargin ? this.maxMargin : margin
 
     if (margin == this.margin) return
-    var { name } = this.props
 
     if (!this.props.hasStickyPosition) {
       this.label.style.top = `${margin}px`
@@ -69,22 +93,27 @@ export default class Category extends React.Component {
   }
 
   getEmojis() {
-    var { name, emojis, perLine } = this.props
+    var { name, emojis, recent, perLine } = this.props
 
     if (name == 'Recent') {
       let { custom } = this.props
-      let frequentlyUsed = frequently.get(perLine)
+      let frequentlyUsed = recent || frequently.get(perLine)
 
       if (frequentlyUsed.length) {
-        emojis = frequentlyUsed.map((id) => {
-          for (let emoji of custom) {
-            if (emoji.id === id) {
+        emojis = frequentlyUsed
+          .map(id => {
+            const emoji = custom.filter(e => e.id === id)[0]
+            if (emoji) {
               return emoji
             }
-          }
 
-          return id
-        })
+            return id
+          })
+          .filter(id => !!getData(id))
+      }
+
+      if (emojis.length === 0 && frequentlyUsed.length > 0) {
+        return null
       }
     }
 
@@ -105,12 +134,20 @@ export default class Category extends React.Component {
     this.container.style.display = display
   }
 
+  setContainerRef(c) {
+    this.container = c
+  }
+
+  setLabelRef(c) {
+    this.label = c
+  }
+
   render() {
-    var { name, hasStickyPosition, emojiProps, i18n } = this.props,
-        emojis = this.getEmojis(),
-        labelStyles = {},
-        labelSpanStyles = {},
-        containerStyles = {}
+    var { id, name, hasStickyPosition, emojiProps, i18n } = this.props,
+      emojis = this.getEmojis(),
+      labelStyles = {},
+      labelSpanStyles = {},
+      containerStyles = {}
 
     if (!emojis) {
       containerStyles = {
@@ -128,37 +165,45 @@ export default class Category extends React.Component {
       }
     }
 
-    return <div ref='container' className={`emoji-mart-category ${emojis && !emojis.length ? 'emoji-mart-no-results' : ''}`} style={containerStyles}>
-      <div style={labelStyles} data-name={name} className='emoji-mart-category-label'>
-        <span style={labelSpanStyles} ref='label'>{i18n.categories[name.toLowerCase()]}</span>
-      </div>
-
-      {emojis && emojis.map((emoji) =>
-        Emoji({
-          emoji: emoji,
-          ...emojiProps
-        })
-      )}
-
-      {emojis && !emojis.length &&
-        <div>
-          <div>
-          {Emoji({
-            ...emojiProps,
-            size: 38,
-            emoji: 'sleuth_or_spy',
-            onOver: null,
-            onLeave: null,
-            onClick: null,
-          })}
-          </div>
-
-          <div className='emoji-mart-no-results-label'>
-            {i18n.notfound}
-          </div>
+    return (
+      <div
+        ref={this.setContainerRef}
+        className={`emoji-mart-category ${emojis && !emojis.length
+          ? 'emoji-mart-no-results'
+          : ''}`}
+        style={containerStyles}
+      >
+        <div
+          style={labelStyles}
+          data-name={name}
+          className="emoji-mart-category-label"
+        >
+          <span style={labelSpanStyles} ref={this.setLabelRef}>
+            {i18n.categories[id]}
+          </span>
         </div>
-      }
-    </div>
+
+        {emojis && emojis.map(emoji => Emoji({ emoji: emoji, ...emojiProps }))}
+
+        {emojis &&
+          !emojis.length && (
+            <div>
+              <div>
+                {Emoji({
+                  ...emojiProps,
+                  size: 38,
+                  emoji: 'sleuth_or_spy',
+                  onOver: null,
+                  onLeave: null,
+                  onClick: null,
+                })}
+              </div>
+
+              <div className="emoji-mart-no-results-label">{i18n.notfound}</div>
+            </div>
+          )}
+      </div>
+    )
   }
 }
 
@@ -169,6 +214,7 @@ Category.propTypes = {
   native: PropTypes.bool.isRequired,
   perLine: PropTypes.number.isRequired,
   emojiProps: PropTypes.object.isRequired,
+  recent: PropTypes.arrayOf(PropTypes.string),
 }
 
 Category.defaultProps = {
