@@ -2,36 +2,50 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import emojiIndex from '../utils/emoji-index'
 
-export default class Search extends React.Component {
-  handleChange() {
-    var { input } = this.refs,
-        value = input.value
+export default class Search extends React.PureComponent {
+  constructor(props) {
+    super(props)
 
-    this.props.onSearch(emojiIndex.search(value, {
-      emojisToShowFilter: this.props.emojisToShowFilter,
-      maxResults: this.props.maxResults,
-      include: this.props.include,
-      exclude: this.props.exclude,
-      custom: this.props.custom,
-    }))
+    this.setRef = this.setRef.bind(this)
+    this.handleChange = this.handleChange.bind(this)
+  }
+
+  handleChange() {
+    var value = this.input.value
+
+    this.props.onSearch(
+      emojiIndex.search(value, {
+        emojisToShowFilter: this.props.emojisToShowFilter,
+        maxResults: this.props.maxResults,
+        include: this.props.include,
+        exclude: this.props.exclude,
+        custom: this.props.custom,
+      })
+    )
+  }
+
+  setRef(c) {
+    this.input = c
   }
 
   clear() {
-    this.refs.input.value = ''
+    this.input.value = ''
   }
 
   render() {
     var { i18n, autoFocus } = this.props
 
-    return <div className='emoji-mart-search'>
-      <input
-        ref='input'
-        type='text'
-        onChange={this.handleChange.bind(this)}
-        placeholder={i18n.search}
-        autoFocus={autoFocus}
-      />
-    </div>
+    return (
+      <div className="emoji-mart-search">
+        <input
+          ref={this.setRef}
+          type="text"
+          onChange={this.handleChange}
+          placeholder={i18n.search}
+          autoFocus={autoFocus}
+        />
+      </div>
+    )
   }
 }
 
@@ -43,7 +57,7 @@ Search.propTypes = {
 }
 
 Search.defaultProps = {
-  onSearch: (() => {}),
+  onSearch: () => {},
   maxResults: 75,
   emojisToShowFilter: null,
   autoFocus: false,
