@@ -1,23 +1,32 @@
 import buildSearch from './build-search'
-import data from '../../data'
+import data from '../data'
+import stringFromCodePoint from '../polyfills/stringFromCodePoint'
+
+const _JSON = JSON
 
 const COLONS_REGEX = /^(?:\:([^\:]+)\:)(?:\:skin-tone-(\d)\:)?$/
-const SKINS = [
-  '1F3FA', '1F3FB', '1F3FC',
-  '1F3FD', '1F3FE', '1F3FF',
-]
+const SKINS = ['1F3FA', '1F3FB', '1F3FC', '1F3FD', '1F3FE', '1F3FF']
 
 function unifiedToNative(unified) {
   var unicodes = unified.split('-'),
-      codePoints = unicodes.map((u) => `0x${u}`)
+    codePoints = unicodes.map(u => `0x${u}`)
 
-  return String.fromCodePoint(...codePoints)
+  return stringFromCodePoint.apply(null, codePoints)
 }
 
 function sanitize(emoji) {
-  var { name, short_names, skin_tone, skin_variations, emoticons, unified, custom, imageUrl } = emoji,
-      id = emoji.id || short_names[0],
-      colons = `:${id}:`
+  var {
+      name,
+      short_names,
+      skin_tone,
+      skin_variations,
+      emoticons,
+      unified,
+      custom,
+      imageUrl,
+    } = emoji,
+    id = emoji.id || short_names[0],
+    colons = `:${id}:`
 
   if (custom) {
     return {
@@ -26,7 +35,7 @@ function sanitize(emoji) {
       colons,
       emoticons,
       custom,
-      imageUrl
+      imageUrl,
     }
   }
 
@@ -78,18 +87,9 @@ function getData(_emoji, skin, set) {
 
     if (data.emojis.hasOwnProperty(emoji)) {
       emojiData = data.emojis[emoji]
+    } else {
+      return null
     }
-  } else if (emoji.custom) {
-    emojiData = emoji
-
-    emojiData.search = buildSearch({
-      short_names: emoji.short_names,
-      name: emoji.name,
-      keywords: emoji.keywords,
-      emoticons: emoji.emoticons
-    })
-
-    emojiData.search = emojiData.search.join(',')
   } else if (emoji.id) {
     if (data.short_names.hasOwnProperty(emoji.id)) {
       emoji.id = data.short_names[emoji.id]
@@ -101,14 +101,23 @@ function getData(_emoji, skin, set) {
     }
   }
 
+  if (!Object.keys(emojiData).length) {
+    emojiData = emoji
+    emojiData.custom = true
+
+    if (!emojiData.search) {
+      emojiData.search = buildSearch(emoji)
+    }
+  }
+
   emojiData.emoticons || (emojiData.emoticons = [])
   emojiData.variations || (emojiData.variations = [])
 
   if (emojiData.skin_variations && skin > 1 && set) {
-    emojiData = JSON.parse(JSON.stringify(emojiData))
+    emojiData = JSON.parse(_JSON.stringify(emojiData))
 
     var skinKey = SKINS[skin - 1],
-        variationData = emojiData.skin_variations[skinKey]
+      variationData = emojiData.skin_variations[skinKey]
 
     if (!variationData.variations && emojiData.variations) {
       delete emojiData.variations
@@ -125,23 +134,27 @@ function getData(_emoji, skin, set) {
   }
 
   if (emojiData.variations && emojiData.variations.length) {
-    emojiData = JSON.parse(JSON.stringify(emojiData))
+    emojiData = JSON.parse(_JSON.stringify(emojiData))
     emojiData.unified = emojiData.variations.shift()
   }
 
   return emojiData
 }
 
+function uniq(arr) {
+  return arr.reduce((acc, item) => {
+    if (acc.indexOf(item) === -1) {
+      acc.push(item)
+    }
+    return acc
+  }, [])
+}
+
 function intersect(a, b) {
-  var aSet = new Set(a),
-      bSet = new Set(b),
-      intersection = null
+  const uniqA = uniq(a)
+  const uniqB = uniq(b)
 
-  intersection = new Set(
-    [...aSet].filter(x => bSet.has(x))
-  )
-
-  return Array.from(intersection)
+  return uniqA.filter(item => uniqB.indexOf(item) >= 0)
 }
 
 function deepMerge(a, b) {
@@ -149,7 +162,7 @@ function deepMerge(a, b) {
 
   for (let key in a) {
     let originalValue = a[key],
-        value = originalValue
+      value = originalValue
 
     if (b.hasOwnProperty(key)) {
       value = b[key]
@@ -165,4 +178,30 @@ function deepMerge(a, b) {
   return o
 }
 
-export { getData, getSanitizedData, intersect, deepMerge, unifiedToNative }
+// https://github.com/sonicdoe/measure-scrollbar
+function measureScrollbar() {
+  if (typeof document == 'undefined') return 0
+  const div = document.createElement('div')
+
+  div.style.width = '100px'
+  div.style.height = '100px'
+  div.style.overflow = 'scroll'
+  div.style.position = 'absolute'
+  div.style.top = '-9999px'
+
+  document.body.appendChild(div)
+  const scrollbarWidth = div.offsetWidth - div.clientWidth
+  document.body.removeChild(div)
+
+  return scrollbarWidth
+}
+
+export {
+  getData,
+  getSanitizedData,
+  uniq,
+  intersect,
+  deepMerge,
+  unifiedToNative,
+  measureScrollbar,
+}
