@@ -1,7 +1,23 @@
 var NAMESPACE = 'emoji-mart'
 
-var isLocalStorageSupported = typeof window !== 'undefined' &&
-                              'localStorage' in window
+const _JSON = JSON
+
+var isLocalStorageSupported =
+  typeof window !== 'undefined' && 'localStorage' in window
+
+let getter
+let setter
+
+function setHandlers(handlers) {
+  handlers || (handlers = {})
+
+  getter = handlers.getter
+  setter = handlers.setter
+}
+
+function setNamespace(namespace) {
+  NAMESPACE = namespace
+}
 
 function update(state) {
   for (let key in state) {
@@ -11,28 +27,31 @@ function update(state) {
 }
 
 function set(key, value) {
-  if (!isLocalStorageSupported) return
-  try {
-    window.localStorage[`${NAMESPACE}.${key}`] = JSON.stringify(value)
-  } catch (e) {
+  if (setter) {
+    setter(key, value)
+  } else {
+    if (!isLocalStorageSupported) return
+    try {
+      window.localStorage[`${NAMESPACE}.${key}`] = _JSON.stringify(value)
+    } catch (e) {}
   }
 }
 
 function get(key) {
-  if (!isLocalStorageSupported) return
-  try {
-    var value = window.localStorage[`${NAMESPACE}.${key}`]
-  } catch (e) {
-    return
-  }
+  if (getter) {
+    return getter(key)
+  } else {
+    if (!isLocalStorageSupported) return
+    try {
+      var value = window.localStorage[`${NAMESPACE}.${key}`]
+    } catch (e) {
+      return
+    }
 
-  if (value) {
-    return JSON.parse(value)
+    if (value) {
+      return JSON.parse(value)
+    }
   }
 }
 
-function setNamespace(namespace) {
-  NAMESPACE = namespace
-}
-
-export default { update, set, get, setNamespace }
+export default { update, set, get, setNamespace, setHandlers }
